new_project/requirements.txt
Flask==2.3.3
Jinja2==3.1.2
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0 